==> Simulation_Analyses/00_original_rough_simulations.Rmd <==
---
title: "Declining_population_simulation"
output:
  html_document: default
  pdf_document: default
---

```{r setup, echo=FALSE, message=FALSE}
library(ggplot2)
library(tidyverse)
```


Set number of simulations to whatever we want to run. Create an empty data frame to store the results in.

```{r}
simulations = 1000

compare <- data.frame(simulations=1:simulations,allyears_trend=NA, surveyeddata_trend=NA, length_of_monitoring=NA, neg_trend_whennot=NA)
```

This is where we are actually running the simulation, and then cutting up the results a few different ways. 
```{r}
for(sims in 1:simulations){

N0 <- runif(1, 1,700) # grabs our starting population size from a random uniform distribution
R <- 2
K <- 500

maxtimesteps = 100 # this is the maximum number of time steps

dat <- data.frame(timestep=1:maxtimesteps,popsize=c(N0, rep(NA, 99))) # this sets our initial population size at time step one and creates an otherwise empty data frame to contain the results

PopNow <- N0 # assigns our current population size to the object PopNow

for(i in 2:nrow(dat)) { 
  dat[i,"popsize"] <- dat[(i-1),"popsize"] + dat[(i-1),"popsize"]*R*(1-dat[(i-1),"popsize"]/K)*rnorm(1, mean=1, sd=0.1)    
} 
# pop at time i is pop at time i-1 + popl at time i-1 times R times 1-pop at i-1 divided by K times error term. 

allyears <- lm(data=dat, popsize ~ timestep) # this looks to see if there is a trend in this simulated population. 

compare[sims,"allyears_trend"] <- allyears$coefficients[2] # this pulls out the beta value for that trend

startdate <- round(runif(1, 1, 100)) # this picks a random start date in those 100 time steps

compare[sims,"length_of_monitoring"] <- 100 - startdate # this this figures out how long the monitoring will run for based on the start date

interval <- round(runif(1, 1, 10),0)

surveytimesteps <- seq(startdate, maxtimesteps, by=interval) # this creates time steps every two years from the start date to the end of our timesteps. 

surveyed_data <- dat[surveytimesteps,] # this subsets our population data to include on the timesteps we surveyed

for(row in 1:nrow(surveyed_data)){ 
  surveyed_data[row,"obspopsize"] <- rnorm(1, mean=surveyed_data[row,"popsize"], sd=0.6*surveyed_data[row,"popsize"])
} # this simulations the observation process, saying that the population size we observe is being pulled from a normal distribution around the actual population size with a 0.6*pop size standard deviation. 

surveyedyears <- lm(data=surveyed_data, obspopsize ~ timestep) # this looks for a trend between our observed population size and time. 

compare[sims,"surveyeddata_trend"] <- surveyedyears$coefficients[2] # this pulls out that beta coefficient

compare[sims,"neg_trend_whennot"] <- ifelse(compare[sims,"surveyeddata_trend"]<0&compare[sims,"allyears_trend"]>0,"yes","no")
} # this tells us whether or not we observed a negative trend when one did not actually occur. 

sdat <- surveyed_data %>% gather("poptype","popsize",-timestep) # rearranging data for graphing
```

This graph shows our very last simulated data set. With the actual population, and what we observed. 
```{r}
ggplot(data=sdat, aes(x=timestep, y=popsize, color=poptype))+geom_line()+geom_smooth(method="lm")
```

So is there a change in the actual population over time?

```{r}
summary(lm(data=surveyed_data, popsize ~ timestep))
```

Is there a change in our observed population over time?

```{r}
summary(lm(data=surveyed_data, obspopsize ~ timestep ))
```

This graph is looking at if there is a relaionship between the trend from our observed population size and the actual population size. 
```{r}
ggplot(data=compare, aes(x=surveyeddata_trend, y=allyears_trend))+geom_point()+geom_smooth(method="lm")

summary(lm(data=compare, surveyeddata_trend ~ length_of_monitoring))
```

This looks at whether the trend is positive or negative and how that changes based on how long we have been monitoring. 

```{r}
ggplot(data=compare, aes(x=length_of_monitoring, y=surveyeddata_trend))+geom_point()+geom_smooth(method="lm")
```


how often do we find a negative trend when there is not one 

So 'yes' means we found a negative trend, when there wasn't one. 
```{r}
ggplot(data=compare, aes(x=neg_trend_whennot)) + 
        geom_bar()
```
